+++ src/lib.rs
@@ -13,6 +13,7 @@ mod conformance;
 #[cfg(feature = "ffi-prototype")]
 #[allow(unsafe_code)]
 mod ffi_prototype;
+mod filesystem;
 mod fixture;
 mod module;
 mod native;
@@ -40,6 +41,7 @@ pub use configuration::{Configuration, ConfigurationValue};
 pub use conformance::FixtureRunner;
 #[cfg(feature = "ffi-prototype")]
 pub use ffi_prototype::{FfiPrototypeError, FfiPrototypeModule};
+pub use filesystem::{initialize_filesystem_plugin, register_filesystem_foreign};
 pub use fixture::{FixtureMetadata, FixtureMetadataError, FixtureOutcome};
 pub use module::{ModuleInstance, ModuleLoadError, ModuleLoader, ModuleSource};
 pub use native::{